chore: review Toolbar
