--- libs/catalog/src/components/DomainFilter/DomainFilter.module.scss
@@ -1,11 +1,0 @@
-.overlay {
-  background: var(--cat-domain-overlay-bg, var(--bg-layer-2, #161b2d));
-  border-color: var(
-    --cat-domain-overlay-border,
-    var(--stroke-secondary, #242c42)
-  );
-}
-
-.triggerActive {
-  color: var(--cat-domain-trigger-active, var(--text-accent-primary, #7da4ff));
-}
