+++ libs/catalog/src/components/Toolbar/Toolbar.module.scss
