+++ libs/catalog/src/constants/catalog-defaults.ts
@@ -7,41 +7,3 @@ export const DEFAULT_SORT_OPTIONS: CatalogSortOption[] = [
   { value: CatalogSortKey.Newest, label: 'Newest' },
   { value: CatalogSortKey.NameAZ, label: 'Name A-Z' },
 ];
-
-/** Default maturity stage options for the Maturity dropdown filter. */
-export const DEFAULT_MATURITY_OPTIONS: string[] = [
-  'Experimental',
-  'Beta',
-  'Production',
-  'Deprecated',
-];
-
-/** Default use-case options for the Use Case dropdown filter. */
-export const DEFAULT_USE_CASE_OPTIONS: string[] = [
-  'Code review',
-  'Code generation',
-  'Document processing',
-  'Summarization',
-  'Data extraction',
-  'Question answering',
-  'Image generation',
-  'Content moderation',
-  'Workflow automation',
-  'Research&Analysis',
-  'Translation',
-  'Classification',
-];
-
-/** Default domain options for the Domain dropdown filter. */
-export const DEFAULT_DOMAIN_OPTIONS: string[] = [
-  'Engineering',
-  'Research',
-  'Finance',
-  'Design',
-  'Sales&Marketing',
-  'Customer Support',
-  'HR&People Ops',
-  'Legal&Compliance',
-  'Data&Analytics',
-  'Operations',
-];
