+++ libs/catalog/src/index.ts
@@ -4,45 +4,19 @@ export { CatalogSortKey } from './types/sort';
 export { CatalogViewMode } from './types/view-mode';
 
 // Models
-export type {
-  CatalogItem,
-  FavoriteItem,
-  TreeNode,
-} from './models/catalog-item';
+export type { CatalogItem, FavoriteItem } from './models/catalog-item';
 export type { CatalogProps, CatalogTitles } from './models/catalog-props';
 
-// Constants
-export {
-  DEFAULT_DOMAIN_OPTIONS,
-  DEFAULT_MATURITY_OPTIONS,
-  DEFAULT_SORT_OPTIONS,
-  DEFAULT_USE_CASE_OPTIONS,
-} from './constants/catalog-defaults';
-export { DEFAULT_ALL_FROM_IDS, DEFAULT_FROM_TREE } from './constants/from-tree';
-
 // Utils
-export {
-  filterCatalogItems,
-  getDomainLabel,
-  getFromLabel,
-  getMaturityLabel,
-  getUseCaseLabel,
-} from './utils/catalog-filter';
+export { filterCatalogItems, getFromLabel } from './utils/catalog-filter';
 export { sortCatalogItems } from './utils/catalog-sort';
-export {
-  applyToggle,
-  findNodeLabel,
-  getAllNodeIds,
-  getDescendantIds,
-  getNodeCheckState,
-} from './utils/catalog-tree';
 export { useFavColumns } from './utils/use-fav-columns';
 
 // Components
 export { Catalog } from './components/Catalog/Catalog';
 
-export { CatalogBrowseToolbar } from './components/CatalogBrowseToolbar/CatalogBrowseToolbar';
-export type { CatalogBrowseToolbarProps } from './components/CatalogBrowseToolbar/CatalogBrowseToolbar';
+export { Toolbar } from './components/Toolbar/Toolbar';
+export type { ToolbarProps } from './models/toolbar-props';
 
 export { CatalogCard } from './components/CatalogCardGrid/CatalogCard';
 export type {
@@ -65,9 +39,6 @@ export type { CatalogFavoritesProps } from './components/CatalogFavorites/Catalo
 export { CatalogListView } from './components/CatalogListView/CatalogListView';
 export type { CatalogListViewProps } from './models/list-props';
 
-export { DomainFilter } from './components/DomainFilter/DomainFilter';
-export type { DomainFilterProps } from './components/DomainFilter/DomainFilter';
-
 export { EntityTypeBadge } from './components/EntityTypeBadge/EntityTypeBadge';
 export type { EntityTypeBadgeProps } from './components/EntityTypeBadge/EntityTypeBadge';
 
@@ -80,17 +51,8 @@ export type { FolderPathProps } from './components/FolderPath/FolderPath';
 export { FromFilter } from './components/FromFilter/FromFilter';
 export type { FromFilterProps } from './components/FromFilter/FromFilter';
 
-export { Highlight } from '@epam/ai-dial-chat-shared';
-export type { HighlightProps } from '@epam/ai-dial-chat-shared';
-
-export { MaturityFilter } from './components/MaturityFilter/MaturityFilter';
-export type { MaturityFilterProps } from './components/MaturityFilter/MaturityFilter';
-
 export { PricingTag } from './components/PricingTag/PricingTag';
 export type { PricingTagProps } from './components/PricingTag/PricingTag';
 
 export { TreeCheckboxRow } from './components/TreeCheckboxRow/TreeCheckboxRow';
 export type { TreeCheckboxRowProps } from './components/TreeCheckboxRow/TreeCheckboxRow';
-
-export { UseCaseFilter } from './components/UseCaseFilter/UseCaseFilter';
-export type { UseCaseFilterProps } from './components/UseCaseFilter/UseCaseFilter';
